+++ conf.lua
@@ -1,5 +1,5 @@
 function love.conf(t)
-    gameVersion = "Eclipse"
+    gameVersion = "Eclipse B"
     t.version = "11.5"
     t.window.title = "NoxTD: Update " .. gameVersion
     t.window.fullscreen = true